+++ Anchor.toml
@@ -3,14 +3,14 @@ seeds = false
 skip-lint = false
 
 [programs.localnet]
-hui = "7syDmCTM9ap9zhfH1gwjDJcGD6LyGFGcggh4fsKxzovV"
+hui = "SQZKnHyyYY6TGFDCKXhH6zFtm9PL7htFttgqwyiTrZT"
 
 [registry]
 url = "https://api.apr.dev"
 
 [provider]
 cluster = "localnet"
-wallet = "/Users/linh.tran/.config/solana/id.json"
+wallet = "/Users/phieu/.config/solana/id.json"
 
 [scripts]
 test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
