+++ app/hui-web/contracts/idl/hui.json
@@ -760,5 +760,8 @@
         ]
       }
     }
-  ]
+  ],
+  "metadata": {
+    "address": "SQZKnHyyYY6TGFDCKXhH6zFtm9PL7htFttgqwyiTrZT"
+  }
 }
