+++ app/hui-web/pages/_app.tsx
@@ -2,12 +2,12 @@ import type { AppProps } from "next/app"
 import { PhantomWalletAdapter } from "@solana/wallet-adapter-phantom"
 import {
   ConnectionProvider,
-  useAnchorWallet,
   WalletProvider,
 } from "@solana/wallet-adapter-react"
 import "@solana/wallet-adapter-react-ui/styles.css"
-// import "../styles/globals.css"
 import { WalletModalProvider } from "@solana/wallet-adapter-react-ui"
+import Layout from "@/components/Layout"
+import "../styles/globals.css"
 
 export default function App({ Component, pageProps }: AppProps) {
   const phantomWallet = new PhantomWalletAdapter()
@@ -16,7 +16,9 @@ export default function App({ Component, pageProps }: AppProps) {
     <ConnectionProvider endpoint={process.env.NEXT_PUBLIC_RPC_URL!}>
       <WalletProvider wallets={[phantomWallet]}>
         <WalletModalProvider>
-          <Component {...pageProps} />
+          <Layout key={2}>
+            <Component {...pageProps} />
+          </Layout>
         </WalletModalProvider>
       </WalletProvider>
     </ConnectionProvider>
